honeypot_api/pyproject.toml
[project]
name = "honeypot-api"
version = "0.1.0"
description = "Agentic Honey-Pot for Scam Detection & Intelligence Extraction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "redis>=5.0.1",
    "python-dotenv>=1.0.0",
    "openai>=1.10.0",
    "tiktoken>=0.5.2",
    "google-generativeai>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
